docs: rebaseline upstream sync workflow
