+++ CONTRIBUTING.md
@@ -24,10 +24,13 @@ bun run verify
 
 ## Upstream Sync Rules
 
-- keep compatibility with `upstream/common-v8` unless an explicit fork decision exists
+- keep compatibility with `upstream/main` unless an explicit fork decision exists
 - avoid dependency downgrades during sync work
 - keep fork-specific behavior isolated and documented
-- always run final gate before sync PR: `bun verify`
+- always run final sync guards before sync PR:
+  - `bun run sync:guard:upstream:strict`
+  - `bun run sync:guard:common-v8:strict` (deprecated alias, temporary)
+- always run final gate before sync PR: `bun run verify:fast`
 
 ## Pull Requests
